requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

boost-cpp=1.72.0=h9d3c048_4
hyperopt=0.2.5=pypi_0
hypopt=1.0.9=pypi_0
opencv-python=4.5.3.56=pypi_0
pandas=1.3.3=pypi_0
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pyhd8ed1ab_0
python-gnupg=0.4.7=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.9.0=py3.8_cuda11.1_cudnn8.0.5_0
scikit-learn=0.24.2=py38hdc147b9_0
scipy=1.5.4=pypi_0
seaborn=0.11.2=pypi_0
torchaudio=0.9.0=py38
torchvision=0.10.0=py38_cu111
tornado=6.1=py38h497a2fe_1
tqdm=4.59.0=pypi_0
xgboost=1.4.2=pypi_0
